+++ BookSales/Controllers/BookController.cs
@@ -1,12 +1,17 @@
-﻿using Entities.Models;
+﻿using BookSales.ActionFilter;
+using Entities.Models;
+using Entities.Pagination;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Store.Application.DTOs.BookDtos;
 using Stroe.Services.IService;
+using System.Text.Json;
 
 namespace BookSales.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [ServiceFilter(typeof(LogFilterAttrubute))]
     public class BookController : ControllerBase
     {
         private readonly IServiceManager _serviceManager;
@@ -17,35 +22,59 @@ namespace BookSales.Controllers
         }
 
         [HttpGet("[action]")]
-        public async Task<IActionResult> GetAllBooks()
+
+        public async Task<IActionResult> GetAllBooks([FromQuery] BookPaginationParameters bookPaginationParameters)
         {
-            var response =await _serviceManager.BookService.GetBookAllAsync(false);
-            return Ok(response);
+
+            var pegedResult = await _serviceManager.BookService.GetBookAllAsync(bookPaginationParameters, false);
+
+            Response.Headers.Add("X-Pagination", JsonSerializer.Serialize(pegedResult.metaData));
+
+            return Ok(pegedResult.books);
         }
         [HttpGet("[action]")]
-        public async Task<IActionResult> GetOneBook( int id )
+        public async Task<IActionResult> GetOneBook(int id)
         {
-            var response = await _serviceManager.BookService.GetBookByIdAsync(id,false); 
+
+            var response = await _serviceManager.BookService.GetBookByIdAsync(id, false);
             return Ok(response);
         }
         [HttpPost("[action]")]
-        public async Task<IActionResult> CreateOneBook(Book book)
+        [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))]
+        public async Task<IActionResult> CreateOneBook(BookDto bookDto)
         {
-            var response = await _serviceManager.BookService.AddBookAsync(book);
+            /*
+              if (!ModelState.IsValid)  //burası  artık ihtiyacımızın olmadıgı bir kısım cunku yurarıda  [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))] bu metotda bu kontrol saglanıyor 
+                 {
+                   return UnprocessableEntity(ModelState);
+                 }
+            */
+            var response = await _serviceManager.BookService.AddBookAsync(bookDto);
             return Ok(response);
         }
 
         [HttpPut("[action]")]
-        public async Task<IActionResult> UpdateOneBook(int id,Book book)
+        [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))]
+        public async Task<IActionResult> UpdateOneBook(int id, BookUpdateDto bookDto)
         {
-            var response = await _serviceManager.BookService.UpdateBookAsync(id, book);
+            /*
+             if (!ModelState.IsValid)  //burası  artık ihtiyacımızın olmadıgı bir kısım cunku yurarıda  [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))] bu metotda bu kontrol saglanıyor 
+                 {
+                      return UnprocessableEntity(ModelState);
+                 }
+           */
+            var response = await _serviceManager.BookService.UpdateBookAsync(id, bookDto);
             return Ok(response);
         }
 
         [HttpDelete("[action]")]
         public async Task<IActionResult> DeleteOnebook(int id)
         {
-            var response = await _serviceManager.BookService.DeleteBookByIdAsync(id,true);
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
+            var response = await _serviceManager.BookService.DeleteBookByIdAsync(id, true);
             return Ok(response);
         }
     }
