+++ BookSales/Controllers/CategoryController.cs
@@ -1,6 +1,8 @@
-﻿using Entities.Models;
+﻿using BookSales.ActionFilter;
+using Entities.Models;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Store.Application.DTOs.CategoryDtos;
 using Stroe.Services.IService;
 using System.Security.AccessControl;
 
@@ -8,6 +10,8 @@ namespace BookSales.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    [ServiceFilter(typeof(LogFilterAttrubute))]
+
     public class CategoryController : ControllerBase
     {
         private readonly IServiceManager _serviceManager;
@@ -20,6 +24,10 @@ namespace BookSales.Controllers
         [HttpGet("[action]")]
         public async Task<IActionResult> GetAllCategories()
         {
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
             var response = await _serviceManager.CategoryService.GetCategoryAllAsync(false);
             return Ok(response);
         }
@@ -27,26 +35,48 @@ namespace BookSales.Controllers
         [HttpGet("[action]")]
         public async Task<IActionResult> GetOneCategory(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
             var response = await _serviceManager.CategoryService.GetCategoryByIdAsync( id, false);
             return Ok(response);
         }
 
-        [HttpPost("[action]")]
-        public async Task<IActionResult> AddOneCategory(Category category)
+     
+        [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))]//dogrulama işlemi için yazıldı //Gelen parametre eksik mi?
+                                                                     //Gelen model istenen kurallara uyuyor mu(örneğin[Required], [Range] gibi)?
+                                                                     //Eksiklikler varsa kullanıcıya uygun bir hata yanıtı döner.
+
+        [HttpPost("[action]")]                                                            
+        public async Task<IActionResult> AddOneCategory(CategoryDto categoryDto)
         {
-            var response = await _serviceManager.CategoryService.AddCategoryAsync(category);
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
+            var response = await _serviceManager.CategoryService.AddCategoryAsync(categoryDto);
             return Ok(response);
         }
         [HttpDelete("[action]")]
         public async Task<IActionResult> DeleteOneCategory(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
             var response = await _serviceManager.CategoryService.DeleteCategoryAsync(id,false);
             return Ok(response);
         }
         [HttpPut("[action]")]
-        public async Task<IActionResult> UpdateOneCategory(int id,Category category)
+        [ServiceFilter(typeof(ValidationModelStateFilterAAttribute))]
+        public async Task<IActionResult> UpdateOneCategory(int id,CategoryDto categoryDto)
         {
-            var response = await _serviceManager.CategoryService.UpdateCategoryAsync(id, category);
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
+            var response = await _serviceManager.CategoryService.UpdateCategoryAsync(id, categoryDto);
             return Ok(response);
         }
     }
